+++ CHANGES.md
@@ -14,6 +14,7 @@
 - MongoDB: Improve UX by using `ctk load table mongodb://...`
 - load table: Refactor to use more OO
 - Add `examples/cloud_import.py`
+- MongoDB: Add adapter for PyMongo
 
 
 ## 2023/11/06 v0.0.2